Trait turbopack::transition::Transition

pub trait Transition {
    fn process_source(&self, asset: AssetVc) -> AssetVc;
    fn process_environment(&self, environment: EnvironmentVc) -> EnvironmentVc;
    fn process_module_options_context(
        &self,
        context: ModuleOptionsContextVc
    ) -> ModuleOptionsContextVc;
    fn process_resolve_options_context(
        &self,
        context: ResolveOptionsContextVc
    ) -> ResolveOptionsContextVc;
    fn process_module(
        &self,
        asset: AssetVc,
        _context: ModuleAssetContextVc
    ) -> AssetVc;
}

Some kind of operation that is executed during reference processing. e. g. you can transition to a different environment on a specific import (reference).

Required Methods

fn process_source(&self, asset: AssetVc) -> AssetVc

Apply modifications/wrapping to the source asset

fn process_environment(&self, environment: EnvironmentVc) -> EnvironmentVc

Apply modifications to the environment

fn process_module_options_context(
&self,
context: ModuleOptionsContextVc
) -> ModuleOptionsContextVc

Apply modifications/wrapping to the module options context

fn process_resolve_options_context(
&self,
context: ResolveOptionsContextVc
) -> ResolveOptionsContextVc

Apply modifications/wrapping to the resolve options context

fn process_module(
&self,
asset: AssetVc,
_context: ModuleAssetContextVc
) -> AssetVc

Apply modifications/wrapping to the final asset