Trait turbopack::transition::Transition

pub trait Transitionwhere
    TaskInput: for<'a> From<&'a Self>,{
    // Provided methods
    fn process_source(&self, asset: AssetVc) -> AssetVc { ... }
    fn process_compile_time_info(
        &self,
        compile_time_info: CompileTimeInfoVc
    ) -> CompileTimeInfoVc { ... }
    fn process_module_options_context(
        &self,
        context: ModuleOptionsContextVc
    ) -> ModuleOptionsContextVc { ... }
    fn process_resolve_options_context(
        &self,
        context: ResolveOptionsContextVc
    ) -> ResolveOptionsContextVc { ... }
    fn process_module(
        &self,
        asset: AssetVc,
        _context: ModuleAssetContextVc
    ) -> AssetVc { ... }
    fn process_context(
        &self,
        context: ModuleAssetContextVc
    ) -> ModuleAssetContextVc { ... }
    fn process(
        &self,
        asset: AssetVc,
        context: ModuleAssetContextVc,
        reference_type: Value<ReferenceType>
    ) -> AssetVc { ... }
}

Some kind of operation that is executed during reference processing. e. g. you can transition to a different environment on a specific import (reference).

Provided Methods

fn process_source(&self, asset: AssetVc) -> AssetVc

Apply modifications/wrapping to the source asset

fn process_compile_time_info( &self, compile_time_info: CompileTimeInfoVc ) -> CompileTimeInfoVc

Apply modifications to the compile-time information

fn process_module_options_context( &self, context: ModuleOptionsContextVc ) -> ModuleOptionsContextVc

Apply modifications/wrapping to the module options context

fn process_resolve_options_context( &self, context: ResolveOptionsContextVc ) -> ResolveOptionsContextVc

Apply modifications/wrapping to the resolve options context

fn process_module( &self, asset: AssetVc, _context: ModuleAssetContextVc ) -> AssetVc

Apply modifications/wrapping to the final asset

fn process_context(&self, context: ModuleAssetContextVc) -> ModuleAssetContextVc

Apply modifications to the context

fn process( &self, asset: AssetVc, context: ModuleAssetContextVc, reference_type: Value<ReferenceType> ) -> AssetVc

Apply modification on the processing of the asset